+++ admin/Settings.php
@@ -2,13 +2,12 @@
 
 namespace TokenToMe\TwitterCards\Admin;
 
-if ( ! function_exists( 'add_action' ) ) {
-	header( 'Status: 403 Forbidden' );
-	header( 'HTTP/1.1 403 Forbidden' );
-	exit();
-}
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+} // Exit if accessed directly
 
 class Settings {
+
 	/**
 	 * settings sections array
 	 *
