+++ composer.json
@@ -18,14 +18,6 @@
   "description": "Easy integration of Twitter cards in WordPress. All card types provided.",
   "require": {
     "composer/installers": "~1.0",
-    "php": ">=5.3.0"
-  },
-  "autoload": {
-    "files": [
-      "functions/functions.php"
-    ],
-    "classmap": [
-      "classes/"
-    ]
+    "php": ">=5.4.0"
   }
 }
